pyproject.toml
[tool.poetry]
name = "certification"
version = "0.1.0"
description = ""
authors = ["KirillOreshchuk <oreshchuk.kirill@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-filter = "^24.2"
flake8 = "^7.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
